origin

+++ src/Foundation/Route.java
@@ -1,11 +1,12 @@
 package Foundation;
 
 import java.io.Serializable;
+import java.net.InetSocketAddress;
 import java.time.LocalDate;
 
 public class Route implements Comparable<Route>, Serializable {
     private static final long serialVersionUID = 1L;
-    private Long id; //Поле не может быть null, Значение поля должно быть больше 0, Значение этого поля должно быть уникальным, Значение этого поля должно генерироваться автоматически
+    private Integer id; //Поле не может быть null, Значение поля должно быть больше 0, Значение этого поля должно быть уникальным, Значение этого поля должно генерироваться автоматически
     private String name; //Поле не может быть null, Строка не может быть пустой
     private Coordinates coordinates; //Поле не может быть null
     private LocalDate creationDate; //Поле не может быть null, Значение этого поля должно генерироваться автоматически
@@ -14,7 +15,7 @@ public class Route implements Comparable<Route>, Serializable {
     private Long distance; //Поле не может быть null, Значение поля должно быть больше 1
     private String login;
 
-    public Route(long id, String name, Coordinates coordinates, Location from, Location to, Long distance, String login) {
+    public Route(int id, String name, Coordinates coordinates, Location from, Location to, Long distance, String login) {
         this.id = id;
         this.name = name;
         this.coordinates = coordinates;
@@ -25,11 +26,11 @@ public class Route implements Comparable<Route>, Serializable {
         this.login = login;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
